# server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ESize": {
				"enum": [
					"small",
					"medium",
					"big"
				],
				"type": "string"
			},
			"Contract": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"longDescription": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"countInStock": {
						"type": "number",
						"format": "double"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"$ref": "#/components/schemas/ESize"
					},
					"sold": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"longDescription",
					"price",
					"countInStock",
					"discount",
					"size",
					"sold",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGetContractsReturn": {
				"properties": {
					"products": {
						"items": {
							"$ref": "#/components/schemas/Contract"
						},
						"type": "array"
					},
					"next": {
						"type": "boolean"
					}
				},
				"required": [
					"products",
					"next"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGetContractsPayload": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"where": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateContractpayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"longDescription": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"countInStock": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"$ref": "#/components/schemas/ESize"
					},
					"sold": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"longDescription",
					"price",
					"countInStock",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"address": {
						"$ref": "#/components/schemas/UserAddress"
					},
					"verified": {
						"type": "boolean"
					},
					"verificationCode": {
						"type": "string",
						"nullable": true
					},
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExp": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"firstName",
					"lastName",
					"email",
					"password",
					"address",
					"verified",
					"verificationCode",
					"resetPasswordToken",
					"resetPasswordExp",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAddress": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"street": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"street",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Profile": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"identityId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"identityId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TPartialUser": {
				"properties": {
					"verified": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"verified",
					"username",
					"email",
					"id"
				],
				"type": "object"
			},
			"TLoginPayload": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"TRegisterPayload": {
				"properties": {
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"confirmPassword",
					"password",
					"username"
				],
				"type": "object"
			},
			"IMessageResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TPasswordResetPayload": {
				"properties": {
					"resetPasswordToken": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"resetPasswordToken",
					"confirmPassword",
					"password"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/contracts": {
			"post": {
				"operationId": "GetContracts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGetContractsReturn"
								}
							}
						}
					}
				},
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGetContractsPayload"
							}
						}
					}
				}
			}
		},
		"/contracts/{id}": {
			"get": {
				"operationId": "GetContract",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Contract"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/contracts/create": {
			"post": {
				"operationId": "CreateContract",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Contract"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateContractpayload"
							}
						}
					}
				}
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/profiles": {
			"post": {
				"operationId": "CreateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Profile"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"facebook": {
										"type": "string"
									},
									"twitter": {
										"type": "string"
									},
									"github": {
										"type": "string"
									},
									"linkedin": {
										"type": "string"
									},
									"avatar": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"facebook",
									"twitter",
									"github",
									"linkedin",
									"avatar"
								]
							}
						}
					}
				}
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Get all users successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Get user with given id successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"Logged in successfully, received user info and jwt token in header": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TPartialUser"
								}
							}
						},
						"headers": {
							"Authorization": {
								"schema": {
									"type": "string"
								},
								"required": true
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TLoginPayload"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "User account created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TPartialUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/users/verify-account/{verificationCode}": {
			"get": {
				"operationId": "VerificateCode",
				"responses": {
					"200": {
						"description": "User account verified succesfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TPartialUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "verificationCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Reset password request successful, email with reset password link sent to the given email",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "ResetPasswordSave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TPasswordResetPayload"
							}
						}
					}
				}
			}
		},
		"/users/reset-password/{resetPasswordToken}": {
			"get": {
				"operationId": "ResetPasswordCheck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "resetPasswordToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/undefined"
		}
	]
}